remove utils.IsValidSlackChannel
