+++ routers/utils/utils.go
@@ -20,25 +20,6 @@ func RemoveUsernameParameterSuffix(name string) string {
 	return name
 }
 
-// IsValidSlackChannel validates a channel name conforms to what slack expects.
-// It makes sure a channel name cannot be empty and invalid ( only an # )
-func IsValidSlackChannel(channelName string) bool {
-	switch len(strings.TrimSpace(channelName)) {
-	case 0:
-		return false
-	case 1:
-		// Keep default behaviour where a channel name is still
-		// valid without an #
-		// But if it contains only an #, it should be regarded as
-		// invalid
-		if channelName[0] == '#' {
-			return false
-		}
-	}
-
-	return true
-}
-
 // SanitizeFlashErrorString will sanitize a flash error string
 func SanitizeFlashErrorString(x string) string {
 	return strings.ReplaceAll(html.EscapeString(x), "\n", "<br>")
