+++ services/forms/repo_form.go
@@ -17,7 +17,6 @@ import (
 	"code.gitea.io/gitea/modules/setting"
 	"code.gitea.io/gitea/modules/structs"
 	"code.gitea.io/gitea/modules/web/middleware"
-	"code.gitea.io/gitea/routers/utils"
 
 	"gitea.com/go-chi/binding"
 )
@@ -305,7 +304,7 @@ type NewSlackHookForm struct {
 // Validate validates the fields
 func (f *NewSlackHookForm) Validate(req *http.Request, errs binding.Errors) binding.Errors {
 	ctx := context.GetContext(req)
-	if f.hasInvalidChannel() {
+	if !isValidSlackChannel(strings.TrimSpace(f.Channel)) {
 		errs = append(errs, binding.Error{
 			FieldNames:     []string{"Channel"},
 			Classification: "",
@@ -315,9 +314,8 @@ func (f *NewSlackHookForm) Validate(req *http.Request, errs binding.Errors) bind
 	return middleware.Validate(errs, ctx.Data, f, ctx.Locale)
 }
 
-// hasInvalidChannel validates the channel name is in the right format
-func (f NewSlackHookForm) hasInvalidChannel() bool {
-	return !utils.IsValidSlackChannel(f.Channel)
+func isValidSlackChannel(name string) bool {
+	return name != "" && name != "#"
 }
 
 // NewDiscordHookForm form for creating discord hook
